+++ RELEASE_NOTES.md
@@ -1,6 +1,22 @@
 Release Notes
 =============
 
+v1.0.0-alpha.X
+--------------
+
+### Breaking changes
+
+- Renamed `TraitBase.isValid` to `isImbued` for symmetry with
+  `imbue`/`imbueTo` methods.
+  [#815](https://github.com/OpenAssetIO/OpenAssetIO/issues/815)
+
+### New features
+
+- Added `TraitBase.isImbuedTo` static/class method, giving a cheaper
+  mechanism for testing whether a `TraitsData` is imbued with a trait.
+  [#815](https://github.com/OpenAssetIO/OpenAssetIO/issues/815)
+
+
 v1.0.0-alpha.9
 --------------